move QueryData to Utilities
